test example: switch from `sprintf()` to safer `snprintf()`
Although the logic of this code guarantees that `sprintf()` will not
overrun its buffer, this commit switches to the safer `snprintf()`
just as a matter of general robustness.

`snprintf()` will never overrun its buffer, even if the calling code
changes (for example if the `CLOCK_DISPLAY_TIME` macro on line 41 changes
to a number larger than 3 digits).